pathPlanning modification done. asearch fails to find destination when there is more than 1 obstacle

+++ MDP_Algo_test/simulation/component.h
@@ -9,7 +9,7 @@ using namespace std;
 // robot 
 const int ROBOT_INIT_Y_GRID = 1;
 const int ROBOT_INIT_X_GRID = 1;
-const float ROBOT_INIT_FACEDIRECTION = 90;
+const int ROBOT_INIT_FACEDIRECTION = 90;
 const float ROBOT_OCCUPY_LENGTH = 22;
 
 // map
